+++ tests/ui/array_get_negative.ml
@@ -1,0 +1,3 @@
+let entrypoint _ =
+  let a = [| 1; 2; 3 |] in
+  a.(-1)